refactor(token): MemberDTO 위치 변경
