+++ backend/src/main/java/wooteco/retrospective/domain/member/Member.java
@@ -1,5 +1,7 @@
 package wooteco.retrospective.domain.member;
 
+import org.apache.commons.lang3.StringUtils;
+
 import java.util.Objects;
 
 public class Member {
@@ -24,7 +26,8 @@ public class Member {
             throw new IllegalArgumentException();
         }
 
-        if (name.length() <= 0 || name.length() > MAX_LENGTH || name.contains(" ")) {
+        if (name.length() <= 0 || name.length() > MAX_LENGTH ||
+                StringUtils.containsWhitespace(name)) {
             throw new IllegalArgumentException();
         }
     }
