+++ backend/src/main/java/wooteco/retrospective/presentation/member/MemberController.java
@@ -5,8 +5,7 @@ import org.springframework.web.bind.annotation.*;
 import wooteco.retrospective.application.dto.MemberLoginDto;
 import wooteco.retrospective.application.member.MemberService;
 import wooteco.retrospective.domain.auth.TokenToName;
-import wooteco.retrospective.domain.dto.MemberDTO;
-import wooteco.retrospective.domain.member.Member;
+import wooteco.retrospective.application.dto.MemberDTO;
 import wooteco.retrospective.presentation.dto.member.MemberLoginRequest;
 import wooteco.retrospective.presentation.dto.member.MemberLoginResponse;
 import wooteco.retrospective.presentation.dto.member.MemberResponse;
